[PlaybackSerialiser] Renamed "integrated" packages to "builtin"
Was always unsatisfied with the name "integrated"...
But I recently took a look at martinvonz/jj and found that they use the name "builtin"
and I just had to switch to it...
