+++ src/main/java/com/minecrafttas/tasmod/TASmodClient.java
@@ -25,12 +25,12 @@ import com.minecrafttas.tasmod.gui.InfoHud;
 import com.minecrafttas.tasmod.handlers.LoadingScreenHandler;
 import com.minecrafttas.tasmod.playback.PlaybackControllerClient;
 import com.minecrafttas.tasmod.playback.PlaybackControllerClient.TASstate;
-import com.minecrafttas.tasmod.playback.filecommands.integrated.DesyncMonitorFileCommandExtension;
-import com.minecrafttas.tasmod.playback.filecommands.integrated.LabelFileCommandExtension;
-import com.minecrafttas.tasmod.playback.filecommands.integrated.OptionsFileCommandExtension;
-import com.minecrafttas.tasmod.playback.metadata.integrated.CreditsMetadataExtension;
-import com.minecrafttas.tasmod.playback.metadata.integrated.StartpositionMetadataExtension;
-import com.minecrafttas.tasmod.playback.tasfile.flavor.integrated.Beta1Flavor;
+import com.minecrafttas.tasmod.playback.filecommands.builtin.DesyncMonitorFileCommandExtension;
+import com.minecrafttas.tasmod.playback.filecommands.builtin.LabelFileCommandExtension;
+import com.minecrafttas.tasmod.playback.filecommands.builtin.OptionsFileCommandExtension;
+import com.minecrafttas.tasmod.playback.metadata.builtin.CreditsMetadataExtension;
+import com.minecrafttas.tasmod.playback.metadata.builtin.StartpositionMetadataExtension;
+import com.minecrafttas.tasmod.playback.tasfile.flavor.builtin.Beta1Flavor;
 import com.minecrafttas.tasmod.registries.TASmodAPIRegistry;
 import com.minecrafttas.tasmod.registries.TASmodConfig;
 import com.minecrafttas.tasmod.registries.TASmodKeybinds;
